requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anthropic==0.25.9
anyio==4.3.0
arxiv==2.1.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.34.105
botocore==1.34.105
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cohere==5.5.0
colorama==0.4.6
curl_cffi==0.7.0b4
dataclasses-json==0.6.6
defusedxml==0.7.1
distro==1.9.0
dnspython==2.6.1
duckduckgo_search==5.3.1
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.3
fastavro==1.9.4
feedparser==6.0.10
filelock==3.14.0
frozenlist==1.4.1
fsspec==2024.5.0
geographiclib==2.0
geojson==2.5.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.3
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.3
google_search_results==2.4.2
googleapis-common-protos==1.63.0
googlemaps==4.10.0
gpsd-py3==0.3.0
greenlet==3.0.3
grpcio==1.63.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.23.0
idna==3.7
imageio==2.34.1
Jinja2==3.1.4
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
langchain==0.1.20
langchain-anthropic==0.1.11
langchain-community==0.0.38
langchain-core==0.1.52
langchain-google-genai==1.0.3
langchain-openai==0.1.6
langchain-text-splitters==0.0.1
langchainhub==0.1.15
langsmith==0.1.57
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
numexpr==2.10.0
numpy==1.26.4
openai==1.28.1
opencv-python==4.9.0.80
orjson==3.10.3
packaging==23.2
pandas==2.2.2
pillow==10.3.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==16.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.1
Pygments==2.18.0
pyowm==3.3.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.10
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
s3transfer==0.10.1
scikit-image==0.23.2
scipy==1.13.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
starlette==0.37.2
streamlit==1.34.0
tenacity==8.3.0
tifffile==2024.5.10
tiktoken==0.6.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
typer==0.12.3
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
watchdog==4.0.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
wikipedia==1.4.0
yarl==1.9.4
youtube-search==2.1.2
